requirements.txt
slack-bolt==1.20.1
slack-sdk==3.27.1
Flask==3.0.3
python-dotenv==1.0.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
gunicorn==22.0.0
